+++ lib/daisy_ui_components.ex
@@ -60,6 +60,7 @@ defmodule DaisyUIComponents do
       import DaisyUIComponents.Input
       import DaisyUIComponents.Join
       import DaisyUIComponents.JSHelpers
+      # import DaisyUIComponents.Label
       import DaisyUIComponents.List
       import DaisyUIComponents.Loading
       import DaisyUIComponents.Menu