+++ CHANGELOG.md
@@ -4,7 +4,15 @@ All notable changes to this project will be documented in this file.
 
 ## [Unreleased]
 
+### Added
+
+- Label component
+- Label storybook
+
+### Changed
+
 - Fix Table ID header
+- Fix CI cache
 
 ## [0.7.6] - 2025-02-17